% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minsample2.R
\name{l_norm}
\alias{l_norm}
\title{Prints the minimum size of the sample required to get epsilon neighborhood for given value of epsilon for Normal Distribution}
\usage{
l_norm(n, eps, mu = 0, sigma = 1)
}
\arguments{
\item{n}{a vector of proposed sample size}

\item{eps}{a vector of the precision level}

\item{mu}{the location parameter for the underlying distribution, here normal distribution(mean)}

\item{sigma}{the scale parameter for the underlying distribution, here normal distribution(standard deviation)}
}
\value{
report: the data frame containing the minimum value of the sample size corresponding to the pre-fixed epsilon
}
\description{
This package helps determining the minimum sample size required to attain some pre-fixed precision level
}
\details{
in any distribution for a large sample the mean-squared error gradually tends to zero, the minimum number depends on the precision level i.e. the pre-fixed eplison
}
\examples{
l_norm(1:5,0.5,3,1)
}
\references{
Methods for this process is described in A.M.Gun,M.K.Gupta,B.Dasgupta(2019,ISBN:81-87567-81-3).
}
